DRY up markdown serving logic
Extract shared utilities to src/utils/docs.ts:
- prefersMarkdown() - Parse Accept header with quality values
- createMarkdownResponse() - Create standardized markdown response

Both doc routes now use these shared functions, eliminating
duplicate code and making maintenance easier.
