+++ src/routes/$libraryId/$version.docs.framework.$framework.$.tsx
@@ -1,31 +1,13 @@
 import { seo } from '~/utils/seo'
 import { Doc } from '~/components/Doc'
-import { loadDocs } from '~/utils/docs'
+import {
+  loadDocs,
+  prefersMarkdown,
+  createMarkdownResponse,
+} from '~/utils/docs'
 import { getBranch, getLibrary } from '~/libraries'
 import { capitalize } from '~/utils/utils'
 import { DocContainer } from '~/components/DocContainer'
-
-// Helper function to check if the Accept header prefers markdown
-function prefersMarkdown(acceptHeader: string | null): boolean {
-  if (!acceptHeader) return false
-
-  const accepts = acceptHeader.split(',').map(type => {
-    const [mediaType, ...params] = type.trim().split(';')
-    const quality = params.find(p => p.trim().startsWith('q='))
-    const q = quality ? parseFloat(quality.split('=')[1]) : 1.0
-    return { mediaType: mediaType.toLowerCase(), q }
-  })
-
-  const markdownQ = accepts.find(a =>
-    a.mediaType === 'text/markdown' || a.mediaType === 'text/plain'
-  )?.q || 0
-
-  const htmlQ = accepts.find(a =>
-    a.mediaType === 'text/html' || a.mediaType === '*/*'
-  )?.q || 0
-
-  return markdownQ > 0 && markdownQ > htmlQ
-}
 
 export const ServerRoute = createServerFileRoute().methods({
   GET: async ({ request, params }) => {
@@ -45,16 +27,7 @@ export const ServerRoute = createServerFileRoute().methods({
         redirectPath: `/${library.id}/${version}/docs/overview`,
       })
 
-      const markdownContent = `# ${doc.title}\n${doc.content}`
-
-      return new Response(markdownContent, {
-        headers: {
-          'Content-Type': 'text/markdown; charset=utf-8',
-          'Cache-Control': 'public, max-age=0, must-revalidate',
-          'Cdn-Cache-Control': 'max-age=300, stale-while-revalidate=300, durable',
-          'Vary': 'Accept',
-        },
-      })
+      return createMarkdownResponse(doc.title, doc.content)
     }
   },
 })
